Implement first round of code reviews for the PR
Use single quotes instead of double quotes in grep part of
SCONTROL_OUTPUT_AWK_PARSER to protect from possible bash
interpretation of the ! character.

+++ src/common/utils.py
@@ -250,7 +250,7 @@ def convert_range_to_list(node_range):
     )
 
 
-def time_is_up(initial_time: datetime, current_time: datetime, grace_time: int):
+def time_is_up(initial_time: datetime, current_time: datetime, grace_time: float):
     """Check if timeout is exceeded."""
     # Localize datetime objects to UTC if not previously localized
     # All timestamps used in this function should be already localized
